move views to Playground.Wpf.UI library, set Playground.Wpf as executable
