+++ Projects/Playground/Playground.Wpf.UI/MainWindow.xaml
@@ -1,0 +1,14 @@
+﻿<views:MvxWindow x:Class="Playground.Wpf.UI.MainWindow"
+                 xmlns:views="clr-namespace:MvvmCross.Platforms.Wpf.Views;assembly=MvvmCross.Platforms.Wpf"
+                 xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                 xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                 xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+                 xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+                 xmlns:local="clr-namespace:Playground.Wpf.UI"
+                 mc:Ignorable="d"
+                 Title="MainWindow"
+                 Height="450"
+                 Width="800">
+    <Grid>
+    </Grid>
+</views:MvxWindow>
