+++ Projects/Playground/Playground.Wpf.UI/MainWindow.xaml.cs
@@ -1,0 +1,29 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
+using System.Windows.Shapes;
+using MvvmCross.Platforms.Wpf.Views;
+
+namespace Playground.Wpf.UI
+{
+    /// <summary>
+    /// Interaction logic for MainWindow.xaml
+    /// </summary>
+    public partial class MainWindow : MvxWindow
+    {
+        public MainWindow()
+        {
+            InitializeComponent();
+        }
+    }
+}
