+++ Projects/Playground/Playground.Wpf.UI/Views/ChildView.xaml
@@ -1,0 +1,25 @@
+﻿<views:MvxWpfView x:Class="Playground.Wpf.UI.Views.ChildView"
+                  x:TypeArguments="vm:ChildViewModel"
+                  xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                  xmlns:views="clr-namespace:MvvmCross.Platforms.Wpf.Views;assembly=MvvmCross.Platforms.Wpf"
+                  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                  xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+                  xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+                  xmlns:local="clr-namespace:Playground.Wpf.UI.Views"
+                  xmlns:vm="clr-namespace:Playground.Core.ViewModels;assembly=Playground.Core"
+                  mc:Ignorable="d"
+                  d:DesignHeight="300"
+                  d:DesignWidth="300">
+    <StackPanel>
+        <TextBlock Text="Child View"
+                   Margin="5" />
+        <Button Content="Show Second Child"
+                Command="{Binding ShowSecondChildCommand}"
+                Margin="5"
+                Padding="5" />
+        <Button Content="Close"
+                Command="{Binding CloseCommand}"
+                Margin="5"
+                Padding="5" />
+    </StackPanel>
+</views:MvxWpfView>
