+++ Projects/Playground/Playground.Wpf.UI/Views/ModalView.xaml.cs
@@ -1,9 +1,9 @@
 ﻿using MvvmCross.Platforms.Wpf.Presenters.Attributes;
 
-namespace Playground.Wpf.Views
+namespace Playground.Wpf.UI.Views
 {
     [MvxWindowPresentation(Identifier = nameof(ModalView), Modal = true)]
-    public partial class ModalView 
+    public partial class ModalView
     {
         public ModalView()
         {
