+++ Projects/Playground/Playground.Wpf.UI/Views/SecondChildView.xaml
@@ -1,16 +1,16 @@
-﻿<views:MvxWpfView x:Class="Playground.WpfCore.Views.NestedModalView"
-                  x:TypeArguments="vm:NestedModalViewModel"
+﻿<views:MvxWpfView x:Class="Playground.Wpf.UI.Views.SecondChildView"
+                  x:TypeArguments="vm:SecondChildViewModel"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:views="clr-namespace:MvvmCross.Platforms.Wpf.Views;assembly=MvvmCross.Platforms.Wpf"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:Playground.WpfCore.Views"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:local="clr-namespace:Playground.Wpf.UI.Views"
              xmlns:vm="clr-namespace:Playground.Core.ViewModels;assembly=Playground.Core"
-             mc:Ignorable="d" 
+             mc:Ignorable="d"
              d:DesignHeight="300" d:DesignWidth="300">
     <StackPanel>
-        <TextBlock Text="Nested Modal" Margin="5" />
+        <TextBlock Text="Second Child" Margin="5" />
         <Button Content="Close" Command="{Binding CloseCommand}" Margin="5" Padding="5" />
     </StackPanel>
 </views:MvxWpfView>
