+++ Projects/Playground/Playground.Wpf.UI/Views/WindowChildView.xaml
@@ -1,13 +1,13 @@
-<views:MvxWpfView x:Class="Playground.Wpf.Views.WindowChildView"
+﻿<views:MvxWpfView x:Class="Playground.Wpf.UI.Views.WindowChildView"
                   x:TypeArguments="vm:WindowChildViewModel"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:views="clr-namespace:MvvmCross.Platforms.Wpf.Views;assembly=MvvmCross.Platforms.Wpf"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:Playground.Wpf.Views"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:local="clr-namespace:Playground.Wpf.UI.Views"
              xmlns:vm="clr-namespace:Playground.Core.ViewModels;assembly=Playground.Core"
-             mc:Ignorable="d" 
+             mc:Ignorable="d"
              d:DesignHeight="300" d:DesignWidth="300">
     <StackPanel Background="LightGray">
         <TextBlock Text="{Binding Text}" Margin="5" />
