+++ Projects/Playground/Playground.Wpf.UI/Views/WindowView.xaml
@@ -1,13 +1,13 @@
-<views:MvxWindow x:Class="Playground.Wpf.Views.WindowView"
+﻿<views:MvxWindow x:Class="Playground.Wpf.UI.Views.WindowView"
                  x:TypeArguments="vm:WindowViewModel"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:views="clr-namespace:MvvmCross.Platforms.Wpf.Views;assembly=MvvmCross.Platforms.Wpf"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:Playground.Wpf.Views"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:local="clr-namespace:Playground.Wpf.UI.Views"
              xmlns:vm="clr-namespace:Playground.Core.ViewModels;assembly=Playground.Core"
-             mc:Ignorable="d" 
+             mc:Ignorable="d"
              d:DesignHeight="300" d:DesignWidth="300"
              Height="300" Width="300"
              Title="{Binding Title}">
@@ -23,7 +23,7 @@
                         <ColumnDefinition />
                         <ColumnDefinition />
                     </Grid.ColumnDefinitions>
-                    <Button Content="Show 1st Child" 
+                    <Button Content="Show 1st Child"
                         Command="{Binding ShowWindowChildCommand}" CommandParameter="1"
                         Margin="5" Padding="5" />
                     <Button Content="Show 2st Child" Grid.Column="1"
