+++ Projects/Playground/Playground.Wpf/App.xaml
@@ -1,8 +1,8 @@
-﻿<views:MvxApplication
-    xmlns:views="clr-namespace:MvvmCross.Platforms.Wpf.Views;assembly=MvvmCross.Platforms.Wpf"
-    x:Class="Playground.Wpf.App"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="clr-namespace:Playground.Wpf"
-    StartupUri="MainWindow.xaml">
+﻿<views:MvxApplication x:Class="Playground.Wpf.App"
+                      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                      xmlns:views="clr-namespace:MvvmCross.Platforms.Wpf.Views;assembly=MvvmCross.Platforms.Wpf"
+                      StartupUri="pack://application:,,,/Playground.Wpf.UI;component/MainWindow.xaml">
+    <Application.Resources>
+    </Application.Resources>
 </views:MvxApplication>
