--- Projects/Playground/Playground.Wpf/MainWindow.xaml
@@ -1,10 +1,0 @@
-﻿<views:MvxWindow x:Class="Playground.Wpf.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Playground.Wpf"
-        xmlns:views="clr-namespace:MvvmCross.Platforms.Wpf.Views;assembly=MvvmCross.Platforms.Wpf"
-        mc:Ignorable="d"
-        Title="MainWindow" Height="350" Width="525">
-</views:MvxWindow>
