--- Projects/Playground/Playground.Wpf/MainWindow.xaml.cs
@@ -1,13 +1,0 @@
-﻿using System.Windows;
-using MvvmCross.Platforms.Wpf.Views;
-
-namespace Playground.Wpf
-{
-    public partial class MainWindow : MvxWindow
-    {
-        public MainWindow()
-        {
-            InitializeComponent();
-        }
-    }
-}
