--- Projects/Playground/Playground.Wpf/Properties/AssemblyInfo.cs
@@ -1,20 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.InteropServices;
-using System.Windows;
-
-[assembly: AssemblyTitle("Playground.Wpf")]
-[assembly: AssemblyDescription("")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("Playground.Wpf")]
-[assembly: AssemblyCopyright("Copyright ©  2017")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-[assembly: ComVisible(false)]
-[assembly: ThemeInfo(
-    ResourceDictionaryLocation.None,
-    ResourceDictionaryLocation.SourceAssembly
-)]
-[assembly: AssemblyVersion("6.5.0.0")]
-[assembly: AssemblyFileVersion("6.5.0.0")]
-[assembly: AssemblyInformationalVersion("6.5.0-remove-framework-android-fragments.1+141.Branch.feature-remove-framework-android-fragments.Sha.cd279b4ed2c240c0b21cb2158e59cfceb6749660")]
