--- Projects/Playground/Playground.Wpf/Views/ChildView.xaml.cs
@@ -1,13 +1,0 @@
-﻿using MvvmCross.Platforms.Wpf.Presenters.Attributes;
-
-namespace Playground.Wpf.Views
-{
-    [MvxContentPresentation]
-    public partial class ChildView
-    {
-        public ChildView()
-        {
-            InitializeComponent();            
-        }
-    }
-}
