--- Projects/Playground/Playground.Wpf/Views/ModalView.xaml
@@ -1,17 +1,0 @@
-﻿<views:MvxWpfView x:Class="Playground.Wpf.Views.ModalView"
-                  x:TypeArguments="vm:ModalViewModel"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:views="clr-namespace:MvvmCross.Platforms.Wpf.Views;assembly=MvvmCross.Platforms.Wpf"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:Playground.Wpf.Views"
-             xmlns:vm="clr-namespace:Playground.Core.ViewModels;assembly=Playground.Core"
-             mc:Ignorable="d" 
-             d:DesignHeight="300" d:DesignWidth="300">
-    <StackPanel>
-        <TextBlock Text="Modal View" Margin="5" />
-        <Button Content="Show Nested Modal" Command="{Binding ShowNestedModalCommand}" Margin="5" Padding="5" />
-        <Button Content="Close" Command="{Binding CloseCommand}" Margin="5" Padding="5" />
-    </StackPanel>
-</views:MvxWpfView>
