--- Projects/Playground/Playground.Wpf/Views/NestedModalView.xaml.cs
@@ -1,13 +1,0 @@
-﻿using MvvmCross.Platforms.Wpf.Presenters.Attributes;
-
-namespace Playground.Wpf.Views
-{
-    [MvxContentPresentation(WindowIdentifier = nameof(ModalView))]
-    public partial class NestedModalView 
-    {
-        public NestedModalView()
-        {
-            InitializeComponent();
-        }
-    }
-}
