--- Projects/Playground/Playground.Wpf/Views/SecondChildView.xaml.cs
@@ -1,14 +1,0 @@
-﻿using MvvmCross.Platforms.Wpf.Presenters.Attributes;
-
-namespace Playground.Wpf.Views
-{
-
-    [MvxContentPresentation]
-    public partial class SecondChildView 
-    {
-        public SecondChildView()
-        {
-            InitializeComponent();
-        }
-    }
-}
