--- Projects/Playground/Playground.Wpf/Views/WindowView.xaml.cs
@@ -1,27 +1,0 @@
-﻿using MvvmCross.Platforms.Wpf.Presenters.Attributes;
-using MvvmCross.Presenters;
-using MvvmCross.Presenters.Attributes;
-using MvvmCross.ViewModels;
-using Playground.Core.ViewModels;
-
-namespace Playground.Wpf.Views
-{
-    public partial class WindowView : IMvxOverridePresentationAttribute
-    {
-        public WindowView()
-        {
-            InitializeComponent();
-        }
-
-        public MvxBasePresentationAttribute PresentationAttribute(MvxViewModelRequest request)
-        {
-            var instanceRequest = request as MvxViewModelInstanceRequest;
-            var viewModel = instanceRequest?.ViewModelInstance as WindowViewModel;
-
-            return new MvxWindowPresentationAttribute
-            {
-                Identifier = $"{nameof(WindowView)}.{viewModel?.Count}"
-            };
-        }
-    }
-}
