--- Projects/Playground/Playground.WpfCore/App.xaml
@@ -1,10 +1,0 @@
-﻿<views:MvxApplication x:Class="Playground.WpfCore.App"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:views="clr-namespace:MvvmCross.Platforms.Wpf.Views;assembly=MvvmCross.Platforms.Wpf"
-             xmlns:local="clr-namespace:Playground.WpfCore"
-             StartupUri="MainWindow.xaml">
-    <Application.Resources>
-         
-    </Application.Resources>
-</views:MvxApplication>
