--- Projects/Playground/Playground.WpfCore/MainWindow.xaml.cs
@@ -1,12 +1,0 @@
-﻿using MvvmCross.Platforms.Wpf.Views;
-
-namespace Playground.WpfCore
-{
-    public partial class MainWindow : MvxWindow
-    {
-        public MainWindow()
-        {
-            InitializeComponent();
-        }
-    }
-}
