--- Projects/Playground/Playground.WpfCore/Setup.cs
@@ -1,34 +1,0 @@
-﻿using Microsoft.Extensions.Logging;
-using MvvmCross.Platforms.Wpf.Core;
-using MvvmCross.Plugin;
-using Serilog;
-using Serilog.Extensions.Logging;
-
-namespace Playground.WpfCore
-{
-    public class Setup : MvxWpfSetup<Core.App>
-    {
-        protected override ILoggerProvider CreateLogProvider()
-        {
-            return new SerilogLoggerProvider();
-        }
-
-        protected override ILoggerFactory CreateLogFactory()
-        {
-            Log.Logger = new LoggerConfiguration()
-                .MinimumLevel.Verbose()
-                .WriteTo.Debug()
-                .CreateLogger();
-
-            return new SerilogLoggerFactory();
-        }
-
-        public override void LoadPlugins(IMvxPluginManager pluginManager)
-        {
-            base.LoadPlugins(pluginManager);
-
-            pluginManager.EnsurePluginLoaded<MvvmCross.Plugin.Messenger.Plugin>();
-            pluginManager.EnsurePluginLoaded<MvvmCross.Plugin.Json.Plugin>();
-        }
-    }
-}
