--- Projects/Playground/Playground.WpfCore/Views/RootView.xaml
@@ -1,15 +1,0 @@
-﻿<views:MvxWpfView  x:Class="Playground.WpfCore.Views.RootView"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:views="clr-namespace:MvvmCross.Platforms.Wpf.Views;assembly=MvvmCross.Platforms.Wpf"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:Playground.WpfCore.Views"
-             mc:Ignorable="d" 
-             d:DesignHeight="300" d:DesignWidth="300">
-    <StackPanel>
-        <Button Content="Show Child (Stack Navigation)" Command="{Binding ShowChildCommand}" Margin="5" Padding="5" />
-        <Button Content="Show New Window (Multiple Windows and Non-stack Navigation)" Command="{Binding ShowWindowCommand}" Margin="5" Padding="5" />
-        <Button Content="Show Modal" Command="{Binding ShowModalCommand}" Margin="5" Padding="5" />
-    </StackPanel>
-</views:MvxWpfView>
