--- Projects/Playground/Playground.WpfCore/Views/RootView.xaml.cs
@@ -1,14 +1,0 @@
-﻿using MvvmCross.ViewModels;
-using Playground.Core.ViewModels;
-
-namespace Playground.WpfCore.Views
-{
-    [MvxViewFor(typeof(RootViewModel))]
-    public partial class RootView
-    {
-        public RootView()
-        {
-            InitializeComponent();
-        }
-    }
-}
