--- Projects/Playground/Playground.WpfCore/Views/WindowView.xaml
@@ -1,37 +1,0 @@
-﻿<views:MvxWindow x:Class="Playground.WpfCore.Views.WindowView"
-                 x:TypeArguments="vm:WindowViewModel"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:views="clr-namespace:MvvmCross.Platforms.Wpf.Views;assembly=MvvmCross.Platforms.Wpf"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:Playground.WpfCore.Views"
-             xmlns:vm="clr-namespace:Playground.Core.ViewModels;assembly=Playground.Core"
-             mc:Ignorable="d" 
-             d:DesignHeight="300" d:DesignWidth="300"
-             Height="300" Width="300"
-             Title="{Binding Title}">
-    <views:MvxWindow.ContentTemplate>
-        <DataTemplate>
-            <Grid>
-                <Grid.RowDefinitions>
-                    <RowDefinition Height="Auto" />
-                    <RowDefinition />
-                </Grid.RowDefinitions>
-                <Grid DataContext="{Binding DataContext, RelativeSource={RelativeSource FindAncestor, AncestorType={x:Type Window}}}">
-                    <Grid.ColumnDefinitions>
-                        <ColumnDefinition />
-                        <ColumnDefinition />
-                    </Grid.ColumnDefinitions>
-                    <Button Content="Show 1st Child" 
-                        Command="{Binding ShowWindowChildCommand}" CommandParameter="1"
-                        Margin="5" Padding="5" />
-                    <Button Content="Show 2st Child" Grid.Column="1"
-                        Command="{Binding ShowWindowChildCommand}" CommandParameter="2"
-                        Margin="5" Padding="5" />
-                </Grid>
-                <ContentPresenter Content="{Binding}" Grid.Row="1" />
-            </Grid>
-        </DataTemplate>
-    </views:MvxWindow.ContentTemplate>
-</views:MvxWindow>
